backend/requirements.txt
# Dependencies for Polla Futbolera Backend
fastapi
uvicorn
gspread
google-auth
requests
pydantic
python-dotenv
